Switched to cmake & made it build on MSVC

+++ src/lib_ym2203/ym2203_emu.cpp
@@ -1353,9 +1353,9 @@ static int init_tables(void)
 		/* we never reach zero here due to ((i*2)+1) */
 
 		if (m>0.0)
-			o = 8*log(1.0/m)/log(2);	/* convert to 'decibels' */
+			o = 8*log(1.0/m)/log((double)2);	/* convert to 'decibels' */
 		else
-			o = 8*log(-1.0/m)/log(2);	/* convert to 'decibels' */
+			o = 8*log(-1.0/m)/log((double)2);	/* convert to 'decibels' */
 
 		o = o / (ENV_STEP/4);
